moved query helpers into utils folder, added config.js to gitignore
