+++ rest-server/src/components/users/userControllers.js
@@ -1,4 +1,4 @@
-import { globalController } from '../../lib/components/';
+import { globalController } from '../../lib/utils/queryHelpers.js';
 import { userQuery } from './userQueries';
 
 export const userController = globalController(userQuery , 'userController');
